feat: IngredientDTO에서 int double로 수정
Resolves #27

+++ backend/src/main/java/com/shinhan/naengtureat/recipe/dto/RecipeIngredientDTO.java
@@ -15,5 +15,5 @@ import lombok.ToString;
 @ToString
 public class RecipeIngredientDTO {
 	private Long ingredientId; // Ingredient ID
-	private int quantity; // 수량
+	private double quantity; // 수량
 }
